LayoutTests/http/tests/workers/text-encoding-expected.txt

This test verifies handling of text encoding in workers. The behavior matches FF3.1b2 with a single exclusion (see below).
This is what's tested:
- If http header 'Content-Type' with 'charset' specified is on response with worker script or XHR payload, that encoding is used.
- In absense of http header, the script of the worker is decoded using the text encoding of the parent document.
- In absense of http header, the content of the XHR request is decoded using UTF-8.
- The URLs used in workers (for subworkers or XHR) are always encoded using UTF-8 (in regular html documents parts of the query an hash may be encoded with other encodings).
- The base URL for the worker (used to resolve relative URLs for subworkers and XHR) is the URL of its script. Note: this is different from FF 3.1b2 which is using the URL of the parent document, which may be a bug (TBD).

Document encoding: windows-1251
Document, Workers: All XHR responses should match this: Привет
Document: : XHR: Привет
Document: Проверка
worker 1: Has no http header with charset
worker 1: Original test string: Привет
worker 1: PASS: XHR query was encoded in UTF-8: Привет
worker 1: Test: Привет (encoded using Windows-1251).
worker 1: Test: рТЙЧЕФ (encoded using koi8-r).
worker 1: XHR: Привет
worker 1: XHR: Привет
worker 1: exit
worker 2: Has http header with charset=koi8-r
worker 2: Original test string: Привет
worker 2: PASS: XHR query was encoded in UTF-8: Привет
worker 2: Test: Привет (encoded using koi8-r).
worker 2: Test: оПХБЕР (encoded using Windows-1251).
worker 2: XHR: Привет
worker 2: XHR: Привет
worker 2: exit
